Move ConsumerConstraint into config package and improve the source-file-path-checks
