+++ satviz-consumer/src/main/java/edu/kit/satviz/consumer/config/ConsumerConstraint.java
@@ -1,15 +1,11 @@
-package edu.kit.satviz.consumer.cli;
+package edu.kit.satviz.consumer.config;
 
 import static edu.kit.satviz.common.FileExistsConstraint.fileExists;
 
 import edu.kit.satviz.common.Constraint;
 import edu.kit.satviz.common.ConstraintValidationException;
-import edu.kit.satviz.consumer.config.ConsumerConfig;
-import edu.kit.satviz.consumer.config.ConsumerMode;
-import edu.kit.satviz.consumer.config.ConsumerModeConfig;
-import edu.kit.satviz.consumer.config.EmbeddedModeConfig;
-import edu.kit.satviz.consumer.config.ExternalModeConfig;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.nio.file.Paths;
 
 /**
@@ -23,19 +19,18 @@ public class ConsumerConstraint implements Constraint<ConsumerConfig> {
       fail("No instance is set");
     }
     fileExists().validate(config.getInstancePath());
-    if (config.getVideoTemplatePath() == null) {
-      fail("No video template path ist set");
-    }
     fileExists().validate(Paths.get(config.getVideoTemplatePath()).getParent());
     ConsumerModeConfig modeConfig = config.getModeConfig();
     if (modeConfig.getMode() == ConsumerMode.EMBEDDED) {
       EmbeddedModeConfig embeddedConfig = (EmbeddedModeConfig) modeConfig;
-      if (embeddedConfig.getSourcePath() == null) {
+      Path source = embeddedConfig.getSourcePath();
+      if (source == null) {
         fail("No source path is set");
-      } else if (Files.isDirectory(embeddedConfig.getSourcePath())) {
-        fail("Directory is set as source");
+      } else if (!Files.isRegularFile(source)) {
+        fail("Source is not a regular file");
+      } else if (!Files.isReadable(source)) {
+        fail("Source is not readable");
       }
-      fileExists().validate(embeddedConfig.getSourcePath());
     } else if (modeConfig.getMode() == ConsumerMode.EXTERNAL) {
       ExternalModeConfig externalConfig = (ExternalModeConfig) modeConfig;
       int port = externalConfig.getPort();
